+++ CriticalMass/Constants.swift
@@ -8,7 +8,6 @@
 import Foundation
 
 enum Constants {
-    static let twitterEndpoint = URL(string: "https://api.criticalmaps.net/twitter")!
     static let apiEndpoint = URL(string: "https://api.criticalmaps.net/")!
     static let criticalMapsiOSGitHubEndpoint = URL(string: "https://github.com/criticalmaps/criticalmaps-ios")!
     static let criticalMapsWebsite = URL(string: "https://www.criticalmaps.net")!