+++ CriticalMass/TwitterManager.swift
@@ -8,31 +8,37 @@
 import Foundation
 
 class TwitterManager {
-    private let url: URL
+    private let request: TwitterRequest
     private var cachedTweets: [Tweet] = [] {
         didSet {
-            updateTweetsCallback?(cachedTweets)
+            contentState = .results(cachedTweets)
+        }
+    }
+
+    private var contentState: ContentState<[Tweet]> = .loading(.default) {
+        didSet {
+            updateContentStateCallback?(contentState)
         }
     }
 
     private let networkLayer: NetworkLayer
 
-    var updateTweetsCallback: (([Tweet]) -> Void)?
+    var updateContentStateCallback: ((ContentState<[Tweet]>) -> Void)?
 
-    init(networkLayer: NetworkLayer, url: URL) {
+    init(networkLayer: NetworkLayer, request: TwitterRequest) {
         self.networkLayer = networkLayer
-        self.url = url
+        self.request = request
     }
 
     public func loadTweets(_ completion: ResultCallback<[Tweet]>? = nil) {
-        let getTweetsRequest = TwitterRequest()
-        networkLayer.get(request: getTweetsRequest) { [weak self] result in
+        networkLayer.get(request: request) { [weak self] result in
             guard let self = self else { return }
             onMain {
                 switch result {
                 case let .failure(error):
-                    completion?(.failure(error))
                     ErrorHandler.default.handleError(error)
+                    self.contentState = .error(.fallback)
+                    completion?(.failure(error))
                 case let .success(response):
                     self.cachedTweets = response.statuses
                     completion?(.success(response.statuses))
@@ -41,10 +47,8 @@ class TwitterManager {
         }
     }
 
-    public func getTweets() -> [Tweet] {
-        if cachedTweets.isEmpty {
-            loadTweets()
-        }
-        return cachedTweets
+    public func getTweets() {
+        contentState = .loading(.default)
+        loadTweets()
     }
 }